requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
aiofiles==23.2.1
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0 
pickle==5.0.1